[Fix]: /admin/info InformationRoute Authorization
